Better View Movement
I have changed the system wherein the view is moved to use boolean values for when the button is pressed and a vector is used to hold the current movement value.

Next I will add a scroll feature enabling the view's size to be changed.

+++ Game Code/Colony Sim/Colony Sim/src/Cells.cpp
@@ -47,8 +47,8 @@ void Cells::m_CreateCellBody(sf::Vector2f dimentions, sf::Vector2f possition)
 */
 void Cells::m_SetGridPos(int x, int y)
 {
-	m_GridPosition.x = x;
-	m_GridPosition.y = y; 
+	m_GridPosition.x = (float)x;
+	m_GridPosition.y = (float)y; 
 }
 
 //--------------------------------------------------------
